add coverage tooltips to preview
to be in sync with list

NHUB-125
